Rename projets for OpsUtils
