+++ OpsUtil.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33516.290
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpsUtil.FileOperationsTests", "OpsUtil.FileOperationsTests\OpsUtil.FileOperationsTests.csproj", "{BB1D253B-0ADD-44F2-8D33-85843D7D1275}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpsUtil.ReplaceInFiles.Cli", "OpsUtil.ReplaceInFiles.Cli\OpsUtil.ReplaceInFiles.Cli.csproj", "{6F7951E4-3567-4B5A-A26E-027762AF90CF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpsUtil.FileOperations", "OpsUtil.FileOperations\OpsUtil.FileOperations.csproj", "{FD4CD571-7F2F-4BDB-8660-8A9737073BAF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BB1D253B-0ADD-44F2-8D33-85843D7D1275}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BB1D253B-0ADD-44F2-8D33-85843D7D1275}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BB1D253B-0ADD-44F2-8D33-85843D7D1275}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BB1D253B-0ADD-44F2-8D33-85843D7D1275}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F7951E4-3567-4B5A-A26E-027762AF90CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F7951E4-3567-4B5A-A26E-027762AF90CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F7951E4-3567-4B5A-A26E-027762AF90CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F7951E4-3567-4B5A-A26E-027762AF90CF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FD4CD571-7F2F-4BDB-8660-8A9737073BAF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FD4CD571-7F2F-4BDB-8660-8A9737073BAF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FD4CD571-7F2F-4BDB-8660-8A9737073BAF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FD4CD571-7F2F-4BDB-8660-8A9737073BAF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8F0890B7-6F23-4996-B506-CC66DB4FFF57}
+	EndGlobalSection
+EndGlobal
