--- ReplaceInFilesTests/Extensions/FileInfoExtensions.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO.Abstractions;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace FinInFilesTests.Extensions
-{
-    public static class FileInfoExtensions
-    {
-        public static string ExtensionWithoutDot(this IFileInfo fileInfo) 
-        { 
-            if(fileInfo == null) 
-                throw new ArgumentNullException(nameof(fileInfo));
-
-            return fileInfo.Extension.Substring(1);
-        }
-
-    }
-}
